// tcfWeb/client/src/app/model/attivita.ts
export class Attivita{
    id_ambito: string;
    id_commessa_cliente: string;
    id_cliente: string;
    id_macro_area: string;
    nome_attivita: string;
    codice_attivita: string;
    descrizione: string;
    budget_euro: number;
    budget_gg: number;
    data_inizio_validita: Date;
    data_fine_validita: Date;
    stato_attivita: string;
}